+++ src/backendResultTransformer.ts
@@ -41,8 +41,18 @@ function setFrameMeta(frame: DataFrame, meta: QueryResultMeta): DataFrame {
 }
 
 function addLevelField(frame: DataFrame, rules: LogLevelRule[]): DataFrame {
+  if (rules.length === 0) {
+    // If there are no rules, we don't need to add the level field
+    return frame;
+  }
   const rows = frame.length ?? frame.fields[0]?.values.length ?? 0;
   const labelsField = frame.fields.find(f => f.name === FrameField.Labels);
+
+  const hasInfoLabel = labelsField?.values.some((value) => value['level']);
+  if (hasInfoLabel) {
+    // If the labels field already has a 'level' label, we don't need to add the level field
+    return frame;
+  }
 
   const levelValues: LogLevel[] = Array.from({ length: rows }, (_, idx) => {
     const labels = (labelsField?.values[idx] ?? {}) as Record<string, any>;